Renomeação do nomes das classes SimpleChat
