+++ JogoDeTruco/nbproject/private/private.xml
@@ -26,20 +26,13 @@
         <group name="github">
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/main/ServidorMain.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Partida.java</file>
+            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Rodada.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Jogada.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/enums/Naipe.java</file>
+            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/comunicacao/ControladorConexao.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/enums/AcaoDaMensagem.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/enums/StatusDaPartida.java</file>
+            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Jogador.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/main/ClienteMain.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/servidor/JogadorListener.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/util/Util.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Jogo.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Rodada.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/enums/Carta.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Mao.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/jogo/Jogador.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/enums/AcaoDaJogada.java</file>
-            <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/enums/EstadoDaMao.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/servidor/Servidor.java</file>
             <file>file:/C:/alencar/RedesComputadoresI/JogoDeTruco/src/comunicacao/Mensagem.java</file>
         </group>
