+++ SimpleChat/nbproject/genfiles.properties
@@ -1,8 +1,8 @@
-build.xml.data.CRC32=dc109b3e
+build.xml.data.CRC32=f8ce19ba
 build.xml.script.CRC32=111d758e
-build.xml.stylesheet.CRC32=8064a381@1.80.1.48
+build.xml.stylesheet.CRC32=8064a381@1.79.1.48
 # This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
 # Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=dc109b3e
-nbproject/build-impl.xml.script.CRC32=e6e07486
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
+nbproject/build-impl.xml.data.CRC32=f8ce19ba
+nbproject/build-impl.xml.script.CRC32=c93696bd
+nbproject/build-impl.xml.stylesheet.CRC32=05530350@1.79.1.48
