+++ SimpleChat/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\alenc\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+user.properties.file=C:\\Users\\alencar.hentges\\AppData\\Roaming\\NetBeans\\8.1\\build.properties
