+++ SimpleChat/nbproject/project.properties
@@ -58,7 +58,7 @@ main.class=main.MainClient
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
-platform.active=default_platform
+platform.active=JDK_1.8
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
