+++ SimpleChat/src/server/Server.java
@@ -1,118 +1,59 @@
 package server;
 
-import classes.Controller;
-import classes.Mensagem;
+import classes.ConnectionController;
 import java.io.IOException;
 import java.io.Serializable;
+import java.net.ServerSocket;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
  * @class Server
  * @author Alencar Rodrigo Hentges <alencarhentges@gmail.com>
  * @date 25/03/2017
  */
-public class Server extends Thread implements Serializable {
+public class Server implements Serializable {
 
-    private static final List<Controller> ALL_CONNECTIONS = new ArrayList<>();
-    private final Controller controler;
+    private final List<ConnectionListener> servers;
+    private final ServerSocket serverSocket;
+    private boolean running;
 
-    public Server(Controller controler) {
-        this.controler = controler;
+    public Server(int port) throws IOException {
+        this.serverSocket = new ServerSocket(port);
+        this.servers = new ArrayList<>();
+        this.running = false;
     }
 
-    /**
-     * Inicia a thread que faz o controle de uma conexão especifica.
-     */
-    @Override
-    public void run() {
-        try {
-            while (this.controler.isConectionOpen()) {
-                Mensagem msg = controler.receber();
-                System.out.println("Servidor recebeu: " + msg);
-                if ("finish_connection".equalsIgnoreCase(msg.getId())) {
-                    this.close();
-                } else {
-                    this.controler.enviar(new Mensagem("ok_connection", null));
-                    enviarMensagemTodos(msg);
-                }
+    public void startServer() throws IOException {
+        if (isRunning()) {
+            return;
+        }
+        this.running = true;
+        System.out.println("Server iniciado!");
+        while (running) {
+            ConnectionController newConnection = new ConnectionController(serverSocket.accept());
+            if (!ConnectionListener.connectionExists(newConnection)) {
+                ConnectionListener connectionTemp = new ConnectionListener(newConnection);
+                connectionTemp.addConnection(newConnection);
+                servers.add(connectionTemp);
+                connectionTemp.start();
             }
-        } catch (IOException ex) {
-            Logger.getLogger(Server.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
-    /**
-     * Envia uma mensagem para todas as conexões existentes exceto para a
-     * conexão que está enviando.
-     *
-     * @param msg
-     */
-    public void enviarMensagemTodos(Mensagem msg) {
-        ALL_CONNECTIONS.forEach((connection) -> {
-            if (connection.isConectionOpen() && !connection.equals(this.controler)) {
-                connection.enviar(new Mensagem(msg.getId(), msg.getMensagem()));
-            }
-        });
-    }
-
-    /**
-     * Adiciona uma lista de conexões na lista de conexões.
-     *
-     * @param newConnections
-     */
-    public void addConnections(List<Controller> newConnections) {
-        newConnections.forEach(connection -> {
-            addConnection(connection);
-        });
-    }
-
-    /**
-     * Adiciona uma nova conexão na lista de conexões.
-     *
-     * @param newConnection
-     */
-    public void addConnection(Controller newConnection) {
-        if (!connectionExists(newConnection)) {
-            ALL_CONNECTIONS.add(newConnection);
+    public void stopServer() throws IOException {
+        if (isRunning()) {
+            this.running = false;
+            this.servers.forEach((server) -> {
+                server.interrupt();
+            });
+            this.serverSocket.close();
+            System.out.println("Server parado!");
         }
     }
 
-    /**
-     * Verifica a existencia da conexão <i>connection</i> na lista de conexões
-     * ativas.
-     *
-     * @param connection
-     * @return <i>true</i> se a conexão existe. <i>false</i> caso a conexão não
-     * exista.
-     */
-    public static boolean connectionExists(Controller connection) {
-        return ALL_CONNECTIONS.contains(connection);
-    }
-
-    /**
-     * Finaliza e remove esta conexão da lista.
-     *
-     * @throws IOException
-     */
-    private void close() throws IOException {
-        ALL_CONNECTIONS.remove(this.controler);
-        this.controler.close(true);
-    }
-
-    /**
-     * Finaliza a conexão. Após isto, faz a interrupção da thread.
-     */
-    @Override
-    public void interrupt() {
-        try {
-            this.close();
-        } catch (IOException ex) {
-            Logger.getLogger(Server.class.getName()).log(Level.SEVERE, null, ex);
-        }
-        super.interrupt();
+    public boolean isRunning() {
+        return this.running;
     }
 
 }
